refactor(core): move the shared Beken driver into beken/driver.rs
run_beken and its flash/erase/read helpers lived in bk7231n.rs, so the
log closure they hand to Transport expanded there too. Every shared-layer
record was therefore tagged tyutool_core::plugins::bk7231n, even when the
job was running on T5AI, T1, T2 or T3:

  [INFO][tyutool_core::plugins::bk7231n] T5AI: starting handshake...

Move the shared driver into the beken module (mirroring esp/common.rs) so
the target names the layer the code actually lives in, and name the chip
in the entry/exit messages instead of relying on the module path. bk7231n.rs
is now a thin plugin shim like t1/t2/t3/t5ai, and the per-chip
"delegating to run_beken" lines are dropped as redundant.
