+++ crates/tyutool-core/src/plugins/beken/mod.rs
@@ -1,7 +1,8 @@
-//! Beken chip protocol layer shared by BK7231N and T5AI plugins.
+//! Beken chip protocol layer shared by the BK7231N, T1, T2, T3 and T5AI plugins.
 
 pub mod chip;
 pub mod command;
+pub mod driver;
 pub mod flash_table;
 pub mod frame;
 pub mod ops;
